drinkable/backend/requirements.txt
fastapi>=0.111.0
uvicorn>=0.29.0

sqlalchemy>=2.0.25
python-jose>=3.3.0
passlib[bcrypt]>=1.7.4

pydantic>=2.7.1

openai>=1.14.0

pillow>=10.2.0
pillow-heif>=0.17.0


